back_end/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==1.4.53
PyJWT==2.8.0
cryptography==41.0.4
PyMySQL==1.1.0

# 이미지 처리
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3

# HTTP 요청
requests==2.31.0

# 환경 변수 관리
python-dotenv==1.0.0

# 개발 도구
pytest==7.4.2
pytest-flask==1.2.0 